Firmware 0.10.0: WiFiManager replaced by the driver's own join and a setup page of ours
The join is the driver's, started and never waited for: the capture, the
log and the console run through it, and a boot with the network away no
longer holds the loop task for 20 s. The setup network is an open soft AP
beside the station, up with no credentials, with the BOOT button held at
power-up, on a refused key, or after ten minutes with no join; it hands
out the board as its DNS and answers a phone's connectivity probe with
the setup page, so the page opens by itself like a hotspot's sign-in
sheet, and the sheet closes when the setup network goes twenty seconds
after the join. The page takes the network and, under an Advanced
revealer, the settings; it focuses the network field, moves to the
password on Return, greys Apply behind a spinner, reports the join as it
happens and names a wrong password, and links the board's address. While
the setup network is up the station keeps still, since its attempts drag
the setup network onto the other channel and a phone mid-join gives up:
a key that has joined before is tried every five minutes, a key that
never joined is erased on its first refusal, and the board's own leaving
is not a verdict. A new key is set only after the attempt in flight is
cancelled, which the driver otherwise refuses. The setup-password
setting is gone with the closed network: the setup network is up for
minutes, on a bike, and goes down at the join.

Every page pins its body to the viewport, follows the system colour
scheme with the scheme declared before the stylesheet, and renders
unset-clock stamps as seconds after boot. The command page polls on tab
selection with the spinner in the output, knows when the MBB is asleep,
and has no button. The poller's last batch is written to its own file at
the asleep edge and read at boot, so the command page has the bike's
last known state, with its age, across reboots; the store neither lists
nor reclaims the board's own files.

Removed with the library: the authentication-failure counting, the
auth-portal window, the no-retry-while-a-portal-is-up rule, the portal
stop's one-second stall, the library's own routes on the setup network,
and its blocking connects.

Flash 1234805 to 1186049 bytes, static RAM 93212 to 92324, no library;
2525 to 2678 lines, the WiFi owner 230 to 256 with the setup network's
rules inside it, the HTTP owner up 81 for the pages, the poller up 47
for the saved batch. Proven from a phone through wrong and right
passwords over some fifteen rounds, with the board flashed over USB while
off the network.

+++ firmware/src/poller.h
@@ -12,7 +12,7 @@ bool pollerRequest();   // false when the MBB has announced its hibernation: the
 bool pollerActive();
 String pollerListJson();
 const String* pollerOutput(const char* name);
-uint32_t pollerOutputAgeS(const char* name);
+long pollerOutputAgeS(const char* name);   // seconds; -1 none, -2 from before this boot with the clock not yet set
 long pollerSoc();              // -1 until known
 bool pollerPack(long& soc, long& packMv, long& currentMa, long& capacityAh, long& tempHiC, long& tempLoC);   // from the last status
 const char* pollerBikeState(); // "" until known
